Refactor E-Series FlashCache module to use module_utils
Refactor the NetApp E-Series module to utlize the common module_utils
and doc_fragments.
